SSTTP 1191 - Correctly calculate future instalment interest
When calculating future interest, the initial payment was not being correctly
subtracted from the oldest debt. Now, the debts are sorted by their due date
so the old debt has the initial payment applied to it first.

Calculation of interest on initial payment has been refactored to use an immutable
and recursive functional approach. Comments have also been added throughout the calculator
service to explain what is going on due to the large amount of specific logic.